requirements.txt
websocket-client~=1.7.0
Pillow~=10.1.0
pynput~=1.7.6
tqdm~=4.66.1